Refactor translation system to be reusable
